lambdas/yaralambda/requirements.txt
yara-python==3.7.0
pymysql==0.7.11
md5checker==0.2.1
